src/python/pyproject.toml
[project]
name = "shoopdaloop"
version = "0.1.1"
description = "Live looping application with DAW elements."
readme = "README.md"
license = { "file" = "LICENSE" }
authors = [
  { name = "Sander Vocke", email = "sandervocke@gmail.com" },
]
maintainers = [
  { name = "Sander Vocke", email = "sandervocke@gmail.com" },
]
requires-python = ">=3.9"

dependencies = [
  "soundfile==0.12.1",
  "mido==1.3.2",
  "jsonschema==4.22.0",
  "numpy==2.1.0",
  "pyjacklib==0.1.1",
  "pyside6-essentials==6.6.3.1",
  "lupa==2.2",
  "appdirs==1.4.4",
  "sounddevice==0.5.0",
  "pywin32 ; platform_system == \"Windows\"",
]

keywords = ['music', 'midi', 'jack', 'looping']

classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Sound/Audio :: Capture/Recording",
  "Operating System :: POSIX :: Linux",
]

[project.gui-scripts]
shoopdaloop = "shoopdaloop.__main__:main"

[project.urls]
Homepage = "https://github.com/SanderVocke/shoopdaloop"
# Documentation = "https://shoopdaloop.readthedocs.io/"
"Bug Tracker" = "https://github.com/SanderVocke/shoopdaloop/issues"
Discussions = "https://github.com/SanderVocke/shoopdaloop/discussions"

[project.optional-dependencies]
test = [
  "pytest",
]
docs = [
  "sphinx",
  "sphinx-material",
  "sphinxcontrib-fulltoc",
  "sphinxcontrib-plantuml"
]

[tool.setuptools.packages.find]
# All the following settings are optional:
where = ["."]
include = ["*"]
exclude = ['codegen', 'dist', 'third_party']

[build-system]
requires = [
  "setuptools",
  "setuptools-scm",
  "ctypesgen@git+https://github.com/SanderVocke/ctypesgen@9b88a970cb85823de83423add2a4b21ea79b63d9",
  "meson",
  "toml"
]